+++ src/vue/VueSatellite.java
@@ -25,8 +25,8 @@ public class VueSatellite extends VueElementMobile {
 	@Override
 	public void redessine() {
 		if (this.satellite.getHautDroit() != null) {
-			int x = this.satellite.getHautDroit().getX() - (this.getImage().getWidth(null) / 2);
-			int y = this.satellite.getBasGauche().getY() - (this.getImage().getHeight(null) / 2);
+			int x = (int) (this.satellite.getHautDroit().getX() - (this.getImage().getWidth(null) / 2));
+			int y = (int) (this.satellite.getBasGauche().getY() - (this.getImage().getHeight(null) / 2));
 			try {
 				this.setCenter(new Point(x, y));
 			} catch (Exception e) {
